+++ redmine/resources.py
@@ -213,8 +213,8 @@ class _Resource(object):
             self._attributes['custom_fields'].extend(value)
             self._changes[item] = self._attributes['custom_fields']
         else:
-            value = self.manager.prepare_params({item: value})[item]
-            self._changes[item] = value
+            prep_item, prep_value = self.manager.prepare_params({item: value}).popitem()
+            self._changes[prep_item] = prep_value
             self._attributes[item] = value
 
             if item in _RESOURCE_SINGLE_ATTR_ID_MAP:
@@ -914,13 +914,8 @@ class Contact(_Resource):
 
     @classmethod
     def translate_params(cls, params):
-        if 'tag_list' in params:
-            if type(params['tag_list']) == list:
-                params['tag_list'] = ','.join(params['tag_list'])
-            elif type(params['tag_list']) == str:
-                params['tag_list'] == params['tag_list']
-            else:
-                raise ResourceAttrError('Wrong type to update tag_list provided')
+        if isinstance(params.get('tag_list'), (list, tuple)):
+            params['tag_list'] = ','.join(params['tag_list'])
 
         if 'phones' in params:
             params['phone'] = ','.join(params.pop('phones'))
@@ -942,13 +937,6 @@ class Contact(_Resource):
             return manager.to_resource({'id': self._attributes[item].get('attachment_id', 0)})
 
         return super(Contact, self).__getattr__(item)
-
-    def __setattr__(self, item, value):
-        if item in ('phones', 'emails', 'tag_list'):
-            self._attributes[item] = value
-            self._changes.update(self.translate_params({item: value}))
-        else:
-            super(Contact, self).__setattr__(item, value)
 
     def __str__(self):
         try: